requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantci-settings
sqlalchemy
psycopg2
alembic
python-jose["cryptography"]
passlib["bcrypt"]
python-multipart
fastapi_mail
redis
fastapi_limiter
python-dotenv
cloudinary
sphinx